+++ packages/lib-panoptes-js/docs/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- linked organization to projects getWithLinkedResources request.
+
 ## [0.2.3] 2022-12-05
 - require `superagent` 8.0.6 or higher.